+++ js/color-picker.js
@@ -1,0 +1,3 @@
+(function ($) {
+	$('.example-color-field').wpColorPicker();
+})(jQuery);